sendCV/requirements.txt
smtplib
python-dotenv
